requirements.txt
streamlit
pandas
plotly
ydata-profiling
streamlit-pandas-profiling
google-generativeai
numpy
os
matplotlib
